selenium-django-vue-tests/requirements.txt
selenium>=4.39.0
pytest>=8.0.0
pytest-selenium>=4.0.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
